refactor: extract creator hub logic
